Move "Test\ConfigurableKernelTestCase" as "KernelTestCase" to the namespace root
