+++ tests/Functional/CompilerPassTest.php
@@ -3,14 +3,14 @@ declare(strict_types=1);
 
 namespace Neusta\Pimcore\TestingFramework\Tests\Functional;
 
+use Neusta\Pimcore\TestingFramework\KernelTestCase;
 use Neusta\Pimcore\TestingFramework\Test\Attribute\RegisterCompilerPass;
-use Neusta\Pimcore\TestingFramework\Test\ConfigurableKernelTestCase;
 use Neusta\Pimcore\TestingFramework\TestKernel;
 use Neusta\Pimcore\TestingFramework\Tests\Fixtures\ConfigurationBundle\DependencyInjection\Compiler\DeregisterSomethingPass;
 use Neusta\Pimcore\TestingFramework\Tests\Fixtures\ConfigurationBundle\DependencyInjection\Compiler\RegisterSomethingPass;
 use Symfony\Component\DependencyInjection\Compiler\PassConfig;
 
-final class CompilerPassTest extends ConfigurableKernelTestCase
+final class CompilerPassTest extends KernelTestCase
 {
     /**
      * @test
